analsysis: rolling window: forward-fill wdw half
For data symmetry reasons it makes a lot of sense,
especially for wide windows, to plot the aggregate
at the window center. We're doing that, see
previous corrections / commits. For the last few
days up to now / today, it then makes sense to
forward-fill the last data point. This is what
this current commit does.