bk: fix build merge with cache
Cache had sort order newest -> oldest
New featch had sort order newest -> oldest
Merge was done by list.extend() which
inserts at the end of the list.
This broke the sort order, and the lookup
of the newest build (from pos 0 of cached
builds). This in turn led to an accumulation
of _duplicate_ builds in the cache. And this
had huge impact on the build rate analysis.